% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormEvaluation.R
\name{EvaPRSD}
\alias{EvaPRSD}
\title{Calculate the pooled RSD for normalization evaluation.}
\usage{
EvaPRSD(FeatureTable, GroupNames, SampleInCol = TRUE, output = FALSE)
}
\arguments{
\item{FeatureTable}{Feature intensity table with samples in column and features in row (default).}

\item{GroupNames}{A character vector indicating the names of each group.}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FLASE} if samples are in row.}

\item{output}{\code{TRUE} will output the result table in current working directory}
}
\value{
This function will return a vector that contains the calculated PRSD for each feature.
}
\description{
Calculate the pooled RSD for normalization evaluation.
}
\examples{
vecPRSD = EvaPRSD(TestingData, GroupNames=c("HY", "SX", "SW", "YC"))
}
