% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{High-quality feature selection}
\usage{
FeatureSelection(
  FeatureTable,
  BlankFilter = 2,
  RtRange = c(0, 100),
  QCRSD = 0.25,
  SQCcor = 0.9,
  IntThreshold = 0,
  SampleInCol = TRUE,
  output = FALSE
)
}
\arguments{
\item{FeatureTable}{Data frame with features in row and samples in column (default).}

\item{BlankFilter}{High-quality when mean(sample intensities) > mean(blank intensities) * \code{BlankFilter}}

\item{RtRange}{Range of the defined retention time window, in minute.}

\item{QCRSD}{Relative standard deviation threshold for QC samples.}

\item{SQCcor}{Pearson's correlation threshold for serial QC samples (recommend: 0.8-0.9).}

\item{IntThreshold}{Feature intensity threshold. Feature is detected when its intensity larger than this value.}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FALSE} if samples are in row.}

\item{output}{\code{TRUE} will output the result table in current working directory}
}
\value{
This function will return the original data frame with an extra column
named "Quality" to indicate the feature quality.
}
\description{
Select high-quality features for quantitative analysis.
}
\details{
\code{FeatureTable} contains measured signal intensities of metabolic features,
with features in row and samples in column (default). The column names should
be sample names, and the first row should be sample group names (e.g. control, case).\cr
The first column should be unique feature identifiers.
For group names, please use: \cr
"RT" for retention time column; \cr
"QC" for quality control samples between real samples (normal QC samples); \cr
"blank" for blank samples; \cr
"SQC_###" for serial QC samples with a certain loading amount.
For example, SQC_1.0 means a serial QC sample with injection volume of 1.0 uL. \cr
An example of \code{FeatureTable} is provided as \code{TestingData} in this package.
}
\examples{
selectedTable = FeatureSelection(TestingData)
}
\references{
Yu, Huaxu, and Tao Huan. "MAFFIN: Metabolomics Sample Normalization
Using Maximal Density Fold Change with High-Quality Metabolic Features and Corrected
Signal Intensities." \emph{bioRxiv} (2021).
}
