% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAFFINNorm.R
\name{MAFFINNorm}
\alias{MAFFINNorm}
\title{MAFFFIN normalization}
\usage{
MAFFINNorm(
  FeatureTable,
  BlankFilter = 2,
  RtRange = c(0, 100),
  QCRSD = 0.25,
  SQCcor = 0.9,
  IntThreshold = 0,
  LR_QC_points = 5,
  QR_QC_points = 7,
  SampleInCol = TRUE,
  output = FALSE
)
}
\arguments{
\item{FeatureTable}{Data frame with features in row and samples in column (default).}

\item{BlankFilter}{A numeric value. High-quality when mean(sample intensities) > mean(blank intensities) * \code{BlankFilter}.}

\item{RtRange}{A numeric vector indicating the range of the defined retention time window, in minute.}

\item{QCRSD}{A numeric value indicating the relative standard deviation threshold for QC samples.}

\item{SQCcor}{A numeric value indicating the Pearson's correlation threshold for serial QC samples (recommend: 0.8-0.9).}

\item{IntThreshold}{A numeric value indicating the feature intensity threshold. Feature is detected when its intensity larger than this value.}

\item{LR_QC_points}{Minimum serial QC data points for quadratic regression.}

\item{QR_QC_points}{Minimum serial QC data points for cubic regression.}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FALSE} if samples are in row.}

\item{output}{\code{TRUE} will output the result table in current working directory.}
}
\value{
This function will return a list that contains four items if \code{RunEvaluation = TRUE}:
the normalized feature table, normalization factors, PRMAD of original data,
and PRMAD of normalized data. The last two items will not be generated if
\code{RunEvaluation = FALSE}
}
\description{
Perform sample normalization using MAFFIN algorithm. MAFFIN algorithm consists
of three modules: high-quality feature selection, MS signal intensity correction,
and maximal density fold change normalization.
}
\details{
\code{FeatureTable} contains measured signal intensities of metabolic features,
with features in row and samples in column (default). The column names should
be sample names, and the first row should be sample group names (e.g. control, case).\cr
The first column should be unique feature identifiers.
For group names, please use: \cr
"RT" for retention time column; \cr
"QC" for quality control samples between real samples (normal QC samples); \cr
"blank" for blank samples; \cr
"SQC_###" for serial QC samples with a certain loading amount.
For example, SQC_1.0 means a serial QC sample with injection volume of 1.0 uL. \cr
An example of \code{FeatureTable} is provided as \code{TestingData} in this package.
}
\examples{
MAFFINTable = MAFFINNorm(TestingData)
}
\references{
Yu, Huaxu, and Tao Huan. "MAFFIN: Metabolomics Sample Normalization
Using Maximal Density Fold Change with High-Quality Metabolic Features and Corrected
Signal Intensities." \emph{bioRxiv} (2021).
}
