% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDFCNorm.R
\name{MDFCNorm}
\alias{MDFCNorm}
\title{Maximal density fold change normalization}
\usage{
MDFCNorm(
  FeatureTable,
  IntThreshold = 0,
  SampleInCol = TRUE,
  output = FALSE,
  OutputNormFactors = TRUE
)
}
\arguments{
\item{FeatureTable}{Feature intensity table with samples in column and features in row (default).}

\item{IntThreshold}{Feature intensity threshold. Feature is detected when its intensity larger than this value.}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FLASE} if samples are in row.}

\item{output}{\code{TRUE} will output the result table in current working directory}

\item{OutputNormFactors}{\code{TRUE} will show the normalization factors after normalization}
}
\value{
This function will return a list contains two items: the normalized feature table,
and a set of normalization factors.
}
\description{
Maximal density fold change normalization
}
\examples{
MDFCNormedTable = MDFCNorm(TestingData)
}
