% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{High-quality feature selection}
\usage{
FeatureSelection(
  FeatureTable,
  BlankFilter = 2,
  RtRange = c(0, 100),
  QCRSD = 0.25,
  SQCcor = 0.9,
  SampleInCol = TRUE,
  output = FALSE
)
}
\arguments{
\item{FeatureTable}{Feature intensity table with samples in column and features in row (default).}

\item{BlankFilter}{High-quality when mean(sample intensities) > mean(blank intensities) * \code{BlankFilter}}

\item{RtRange}{Range of the defined retention time window, in minute.}

\item{QCRSD}{Relative standard deviation threshold for QC samples}

\item{SQCcor}{Pearson's correlation threshold for serial QC samples (recommend: 0.8-0.9)}

\item{SampleInCol}{\code{TRUE} if samples are in column.}

\item{output}{\code{TRUE} will output the result table in current working directory}
}
\value{
This function will return the original feature table with an extra column named "Quality" to indicate the feature quality
}
\description{
Select high-quality features for quantitative analysis.
}
\details{
The first row should be sample names, and the second row should be group names.
For group names, please use: "RT" for retention time column,
"QC" for quality control samples between real samples (normal QC samples);
"blank" for blank samples;
"SQC_#amount#" for serial QC samples with a certain loading amount. For example, SQC_1.0 means a serial QC sample with injection volume as 1.0 uL;
}
\examples{
Please see GitHub for demo.
}
