% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntCorrection.R
\name{IntCorrection}
\alias{IntCorrection}
\title{Intensity Correction Using Serial QC Samples}
\usage{
IntCorrection(
  FeatureTable,
  IntThreshold = 0,
  LR_QC_points = 5,
  QR_QC_points = 7,
  SQCcor = 0.9,
  SampleInCol = TRUE,
  output = FALSE
)
}
\arguments{
\item{FeatureTable}{Feature intensity table with samples in column and features in row (default).}

\item{IntThreshold}{Feature intensity threshold. Feature is detected when its intensity larger than this value.}

\item{LR_QC_points}{Required data points for quadratic regression (>= this value)}

\item{QR_QC_points}{Required data points for cubic regression (>= this value)}

\item{SQCcor}{Pearson's correlation threshold for serial QC samples (recommend: 0.8-0.9)}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FLASE} if samples are in row.}

\item{output}{\code{TRUE} will output the result table in current working directory}
}
\value{
This function will return the original feature table with corrected intensities.
}
\description{
Intensity Correction Using Serial QC Samples
}
\examples{
Please see GitHub for demo.
}
