% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAFFINNorm.R
\name{MAFFINNorm}
\alias{MAFFINNorm}
\title{MAFFFIN normalization}
\usage{
MAFFINNorm(
  FeatureTable,
  BlankFilter = 2,
  RtRange = c(0, 100),
  QCRSD = 0.25,
  SQCcor = 0.9,
  IntThreshold = 0,
  LR_QC_points = 5,
  QR_QC_points = 7,
  SampleInCol = TRUE,
  output = FALSE
)
}
\arguments{
\item{FeatureTable}{Feature intensity table with samples in column and features in row (default).}

\item{BlankFilter}{High-quality when mean(sample intensities) > mean(blank intensities) * \code{BlankFilter}.}

\item{RtRange}{Range of the defined retention time window, in minute.}

\item{QCRSD}{Relative standard deviation threshold for QC samples.}

\item{SQCcor}{Pearson's correlation threshold for serial QC samples (recommend: 0.8-0.9).}

\item{IntThreshold}{Feature intensity threshold. Feature is detected when its intensity larger than this value.}

\item{LR_QC_points}{Required data points for quadratic regression (>= this value)}

\item{QR_QC_points}{Required data points for cubic regression (>= this value)}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FLASE} if samples are in row.}

\item{output}{\code{TRUE} will output the result table in current working directory.}
}
\value{
This function will return a list contains two items: the MAFFIN result table,
and a set of normalization factors.
}
\description{
MAFFFIN normalization
}
\examples{
Please see GitHub for demo.
}
