% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDFCNorm.R
\name{MDFCNorm}
\alias{MDFCNorm}
\title{Maximal density fold change normalization}
\usage{
MDFCNorm(
  FeatureTable,
  IntThreshold = 0,
  SampleInCol = TRUE,
  output = FALSE,
  OutputNormFactors = FALSE,
  RunEvaluation = TRUE,
  bwOpt = NA
)
}
\arguments{
\item{FeatureTable}{Data frame with features in row and samples in column (default).}

\item{IntThreshold}{A numeric value indicating the feature intensity threshold. Feature is detected when its intensity larger than this value.}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FALSE} if samples are in row.}

\item{output}{\code{TRUE} will output the result table in current working directory}

\item{OutputNormFactors}{\code{TRUE} will print the normalization factors after normalization}

\item{RunEvaluation}{\code{TRUE} will evaluate the normalization results by intragroup variation.}

\item{bwOpt}{\code{NA} will automatically optimize the bandwidth. Use a numeric value to set the bandwidth and skip the optimization.}
}
\value{
This function will return a list that contains four items if \code{RunEvaluation = TRUE}:
the normalized feature table, normalization factors, PRMAD of original data,
and PRMAD of normalized data. The last two items will not be generated if
\code{RunEvaluation = FALSE}
}
\description{
Sample normalization using maximal density fold change.
}
\details{
\code{FeatureTable} contains measured or corrected signal intensities of metabolic features,
with features in row and samples in column (default). The column names should
be sample names, and the first row should be sample group names (e.g. control, case).\cr
The first column should be unique feature identifiers.
For group names, please use: \cr
"RT" for retention time column; \cr
"QC" for quality control samples between real samples (normal QC samples); \cr
"blank" for blank samples; \cr
"SQC_###" for serial QC samples with a certain loading amount.
For example, SQC_1.0 means a serial QC sample with injection volume of 1.0 uL. \cr
An example of \code{FeatureTable} is provided as \code{TestingData} in this package.
}
\examples{
MDFCNormedTable = MDFCNorm(TestingData)
}
\references{
Yu, Huaxu, and Tao Huan. "MAFFIN: Metabolomics Sample Normalization
Using Maximal Density Fold Change with High-Quality Metabolic Features and Corrected
Signal Intensities." \emph{bioRxiv} (2021).
}
